Other: Move FontSizeEditing to font/fontsize module.
